requirements.txt
platformio
PyQt5
pyserial